feat: address 테이블 변경사항 반영
- id값의 type을 int에서 long으로 변경

+++ bible/src/main/java/com/library/bible/address/respository/IAddressRepository.java
@@ -10,10 +10,10 @@ import com.library.bible.address.model.Address;
 @Mapper
 @Repository
 public interface IAddressRepository {
-    List<Address> selectAddressesByMemId(int memId);
-    Address seleAddress(int addressId);
+    List<Address> selectAddressesByMemId(long memId);
+    Address seleAddress(long addressId);
     int insertAddress(Address address);
     int updateAddress(Address address);
-    int deleteAddressesByMemId(int memId);
-    int deleteAddress(int addressId);
+    int deleteAddressesByMemId(long memId);
+    int deleteAddress(long addressId);
 }
